feat: detach function to useHomeTimeline hook
